+++ src/account/builder.rs
@@ -151,6 +151,7 @@ impl AccountBuilder {
                         // `iota` or `atoi`?
                         None => client
                             .get_bech32_hrp()
+                            .await
                             .unwrap_or_else(|_| SHIMMER_TESTNET_BECH32_HRP.to_string()),
                     }
                 };