+++ src/account/operations/transaction/high_level/burning_melting/burn_nft.rs
@@ -69,7 +69,7 @@ impl AccountHandle {
     // governor address from the nft output.
     async fn output_id_and_basic_output_for_nft(&self, nft_id: NftId) -> crate::Result<(OutputId, Output)> {
         let account = self.read().await;
-        let token_supply = self.client.get_token_supply()?;
+        let token_supply = self.client.get_token_supply().await?;
 
         let (output_id, nft_output) = account
             .unspent_outputs()