+++ src/account/update.rs
@@ -3,7 +3,7 @@
 
 use std::collections::HashMap;
 
-use iota_client::{api_types::response::OutputResponse, block::output::OutputId, Client};
+use iota_client::{api_types::response::OutputWithMetadataResponse, block::output::OutputId, Client};
 
 use crate::account::{
     handle::AccountHandle,
@@ -32,12 +32,12 @@ impl AccountHandle {
         &self,
         addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>,
         unspent_outputs: Vec<OutputData>,
-        spent_or_not_synced_outputs: HashMap<OutputId, Option<OutputResponse>>,
+        spent_or_not_synced_outputs: HashMap<OutputId, Option<OutputWithMetadataResponse>>,
         options: &SyncOptions,
     ) -> crate::Result<()> {
         log::debug!("[SYNC] Update account with new synced transactions");
 
-        let network_id = self.client.get_network_id()?;
+        let network_id = self.client.get_network_id().await?;
         let mut account = self.write().await;
         #[cfg(feature = "events")]
         let account_index = account.index;
@@ -261,7 +261,7 @@ impl AccountHandle {
     // Should only be called from the AccountManager so all accounts are on the same state
     pub(crate) async fn update_account_with_new_client(&mut self, client: Client) -> crate::Result<()> {
         self.client = client;
-        let bech32_hrp = self.client.get_bech32_hrp()?;
+        let bech32_hrp = self.client.get_bech32_hrp().await?;
         log::debug!("[UPDATE ACCOUNT WITH NEW CLIENT] new bech32_hrp: {}", bech32_hrp);
         let mut account = self.write().await;
         for address in &mut account.addresses_with_unspent_outputs {