+++ CHANGELOG.md
@@ -19,6 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.3 - 20XX-XX-XX
+
+### Changed
+
+- Use `OutputWithMetadataResponse` instead of `OutputResponse`;
+- `AccountHandle::build_transaction_essence` made async;
+
 ## 1.0.0-rc.2 - 2022-10-28
 
 ### Added